+++ lib/libesp32/berry/default/be_modtab.c
@@ -393,7 +393,4 @@ BERRY_API void be_load_custom_libs(bvm *vm)
     (void)vm;   /* prevent a compiler warning */
     
     /* add here custom libs */
-#if !BE_USE_PRECOMPILED_OBJECT
-    /* be_load_xxxlib(vm); */
-#endif
 }