+++ lib/libesp32/berry/src/be_strictlib.c
@@ -22,19 +22,11 @@ static int m_init(bvm *vm)
     be_return_nil(vm);
 }
 
-#if !BE_USE_PRECOMPILED_OBJECT
-be_native_module_attr_table(strict) {
-    be_native_module_function("init", m_init),
-};
-
-be_define_native_module(strict, NULL);
-#else
 /* @const_object_info_begin
 module strict (scope: strict, depend: BE_USE_STRICT_MODULE) {
     init, func(m_init)
 }
 @const_object_info_end */
 #include "../generate/be_fixed_strict.h"
-#endif
 
 #endif /* BE_USE_STRICT_MODULE */